Mage Fire
- Added an option to start pull rotation with Combustion, check UI and enable Use combution Early 
- This is not follow any Simc DPS logic and will use combustion as first spell

Why Simc don't use Combustion directly on pull? This is simple, usually we start with 3 charges of Fireblast + 2 charges of Phoanix flame, this is mean that we have 2 pyroblast/flamestrike for free without entring Combustion, the left charge of Fireblast is kept for Combustion

This is being said I advice to don't use this option for big dungeons or Mythic raid, this is up to you.