Add files via upload
**Globals**

Interrupt system: Units with level 82+ are now considered Bosses. This prevents the use of crowd control (CC) on them.

Added a whitelist of NPCs to force the rotation to work on them. Please report any NPC IDs we should DPS but the rotation currently doesn't!

**Beast Mastery Hunter**

Fixed Misdirection casting randomly in certain cases.

Fixed Feign Death usage for dispelling.

Added potion support. Create a macro for your potion and bind it to Universal3.

Cleaned up the UI.